requirements.txt
amqp==2.6.0
asgiref==3.2.10
billiard==3.6.3.0
celery==4.4.6
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
coverage==5.1
cryptography==2.9.2
cycler==0.10.0
decorator==4.4.2
dj-database-url==0.5.0
Django==3.0.7
django-cors-headers==3.4.0
django-extra-fields==2.0.5
django-heroku==0.3.1
django-hitcount==1.3.2
django-redis==4.12.1
django-rest-framework==0.1.0
django-rest-swagger==2.2.0
django-storages==1.9.1
django-taggit==1.3.0
django-taggit-serializer==0.1.7
djangorestframework==3.11.0
drf-jwt==1.16.2
future==0.18.2
gunicorn==20.0.4
idna==2.9
imageio==2.8.0
importlib-metadata==1.6.1
itypes==1.2.0
Jinja2==2.11.2
kiwisolver==1.2.0
kombu==4.6.11
MarkupSafe==1.1.1
matplotlib==3.2.2
networkx==2.4
numpy==1.19.0
openapi-codec==1.3.2
Pillow==7.1.2
psycopg2==2.8.5
psycopg2-binary==2.8.5
pycparser==2.20
PyJWT==1.7.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
PyWavelets==1.1.1
redis==3.5.3
requests==2.24.0
scikit-image==0.17.2
scipy==1.5.0
simplejson==3.17.0
six==1.15.0
sqlparse==0.3.1
tifffile==2020.6.3
uritemplate==3.0.1
urllib3==1.25.9
vine==1.3.0
whitenoise==5.1.0
zipp==3.1.0
